+++ packages/backend-data/src/types.ts
@@ -130,6 +130,7 @@ export type DataProps = {
 };
 
 export type AmplifyDataError =
+  | 'DefineDataConfigurationError'
   | 'InvalidSchemaAuthError'
   | 'InvalidSchemaError'
   | 'MultipleSingletonResourcesError';