+++ packages/deployed-backend-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/deployed-backend-client
 
+## 0.4.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [aec89f9]
+- Updated dependencies [2a69684]
+  - @aws-amplify/platform-core@0.5.0-beta.4
+
 ## 0.4.0-beta.4
 
 ### Minor Changes